move blockchain dependendent validateUncle checks
