+++ packages/blockchain/src/index.ts
@@ -576,7 +576,8 @@ export default class Blockchain implements BlockchainInterface {
       if (this._validateBlocks && !isGenesis) {
         // this calls into `getBlock`, which is why we cannot lock yet
         await this.validateHeader(item instanceof Block ? item.header : item)
-        await block.validate(this, isHeader)
+        await this._validateUncleHeaders(block)
+        await block.validate(isHeader)
       }
 
       if (this._validateConsensus) {
@@ -744,6 +745,94 @@ export default class Blockchain implements BlockchainInterface {
         }
       }
     }
+  }
+
+  /**
+   * The following rules are checked in this method:
+   * Uncle Header is a valid header.
+   * Uncle Header is an orphan, i.e. it is not one of the headers of the canonical chain.
+   * Uncle Header has a parentHash which points to the canonical chain. This parentHash is within the last 7 blocks.
+   * Uncle Header is not already included as uncle in another block.
+   * @param blockchain - pointer to the blockchain
+   */
+  // TODO: Move validateUncleHeaders to blockchain
+  private async _validateUncleHeaders(block: Block) {
+    const uncleHeaders = block.uncleHeaders
+    if (uncleHeaders.length == 0) {
+      return
+    }
+
+    // Each Uncle Header is a valid header
+    await Promise.all(uncleHeaders.map((uh) => this.validateHeader(uh, block.header.number)))
+
+    // Check how many blocks we should get in order to validate the uncle.
+    // In the worst case, we get 8 blocks, in the best case, we only get 1 block.
+    const canonicalBlockMap: Block[] = []
+    let lowestUncleNumber = block.header.number
+
+    uncleHeaders.map((header) => {
+      if (header.number < lowestUncleNumber) {
+        lowestUncleNumber = header.number
+      }
+    })
+
+    // Helper variable: set hash to `true` if hash is part of the canonical chain
+    const canonicalChainHashes: { [key: string]: boolean } = {}
+
+    // Helper variable: set hash to `true` if uncle hash is included in any canonical block
+    const includedUncles: { [key: string]: boolean } = {}
+
+    // Due to the header validation check above, we know that `getBlocks` is between 1 and 8 inclusive.
+    const getBlocks = Number(block.header.number - lowestUncleNumber + BigInt(1))
+
+    // See Geth: https://github.com/ethereum/go-ethereum/blob/b63bffe8202d46ea10ac8c4f441c582642193ac8/consensus/ethash/consensus.go#L207
+    // Here we get the necessary blocks from the chain.
+    let parentHash = block.header.parentHash
+    for (let i = 0; i < getBlocks; i++) {
+      const parentBlock = await this.getBlock(parentHash)
+      if (!parentBlock) {
+        const msg = block._errorMsg('could not find parent block')
+        throw new Error(msg)
+      }
+      canonicalBlockMap.push(parentBlock)
+
+      // mark block hash as part of the canonical chain
+      canonicalChainHashes[parentBlock.hash().toString('hex')] = true
+
+      // for each of the uncles, mark the uncle as included
+      parentBlock.uncleHeaders.map((uh) => {
+        includedUncles[uh.hash().toString('hex')] = true
+      })
+
+      parentHash = parentBlock.header.parentHash
+    }
+
+    // Here we check:
+    // Uncle Header is an orphan, i.e. it is not one of the headers of the canonical chain.
+    // Uncle Header is not already included as uncle in another block.
+    // Uncle Header has a parentHash which points to the canonical chain.
+
+    uncleHeaders.map((uh) => {
+      const uncleHash = uh.hash().toString('hex')
+      const parentHash = uh.parentHash.toString('hex')
+
+      if (!canonicalChainHashes[parentHash]) {
+        const msg = block._errorMsg(
+          'The parent hash of the uncle header is not part of the canonical chain'
+        )
+        throw new Error(msg)
+      }
+
+      if (includedUncles[uncleHash]) {
+        const msg = block._errorMsg('The uncle is already included in the canonical chain')
+        throw new Error(msg)
+      }
+
+      if (canonicalChainHashes[uncleHash]) {
+        const msg = block._errorMsg('The uncle is a canonical block')
+        throw new Error(msg)
+      }
+    })
   }
 
   /**
